services/orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pymongo==4.8.0
python-dotenv==1.0.1
redis==5.0.8
grpcio==1.66.2
crewai>=0.108.0,<1.0.0
crewai-tools>=0.2.4,<1.0.0
requests>=2.32.0
langfuse==2.*